+++ meerket/meerket-infrastructure/src/main/java/org/j1p5/infrastructure/kakao/dto/KakaoProfile.java
@@ -1,9 +1,10 @@
-package org.j1p5.domain.auth.dto;
+package org.j1p5.infrastructure.kakao.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
+import org.j1p5.domain.auth.dto.OauthProfile;
 
-public class KakaoProfile implements OauthProfile{
+public class KakaoProfile implements OauthProfile {
 
     @JsonProperty("id")
     private String id;
