+++ meerket/meerket-infrastructure/src/main/java/org/j1p5/infrastructure/kakao/dto/KakaoToken.java
@@ -1,7 +1,8 @@
-package org.j1p5.domain.auth.dto;
+package org.j1p5.infrastructure.kakao.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
+import org.j1p5.domain.auth.dto.OauthToken;
 
 @Getter
 public class KakaoToken implements OauthToken {
