+++ meerket/meerket-infrastructure/src/main/java/org/j1p5/infrastructure/naver/dto/NaverProfile.java
@@ -1,7 +1,8 @@
-package org.j1p5.domain.auth.dto;
+package org.j1p5.infrastructure.naver.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
+import org.j1p5.domain.auth.dto.OauthProfile;
 
 public class NaverProfile implements OauthProfile {
